refactor: move favorite types/schema to new module
